+++ packages/calcite-components/src/components/list/list.tsx
@@ -219,8 +219,12 @@ export class List
    */
   @property({ reflect: true }) interactionMode: InteractionMode = "interactive";
 
-  /** Specifies an accessible name for the component. */
-  @property() label: string;
+  /**
+   * Specifies an accessible name for the component.
+   *
+   * When `dragEnabled` is `true` and multiple list sorting is enabled with `group`, specifies the component's name for dragging between lists.
+   */
+  @property() label!: string;
 
   /** When `true`, a busy indicator is displayed. */
   @property({ reflect: true }) loading = false;
@@ -685,15 +689,11 @@ export class List
   private updateFilteredItems(): void {
     const { visibleItems, filteredData, filterText } = this;
 
-    const values = filteredData.map((item) => item.value);
-
     const lastDescendantItems = visibleItems?.filter((listItem) =>
       visibleItems.every((li) => li === listItem || !listItem.contains(li)),
     );
 
-    const filteredItems =
-      visibleItems.filter((item) => !filterText || values.includes(item.value)) || [];
-
+    const filteredItems = !filterText ? visibleItems || [] : filteredData.map((item) => item.el);
     const visibleParents = new WeakSet<HTMLElement>();
 
     lastDescendantItems.forEach((listItem) =>
@@ -727,15 +727,23 @@ export class List
     this.updateFilteredData();
   }
 
+  private get effectiveFilterProps(): string[] {
+    if (!this.filterProps) {
+      return ["description", "label", "metadata"];
+    }
+
+    return this.filterProps.filter((prop) => prop !== "el");
+  }
+
   private performFilter(): void {
-    const { filterEl, filterText, filterProps } = this;
+    const { filterEl, filterText, effectiveFilterProps } = this;
 
     if (!filterEl) {
       return;
     }
 
     filterEl.value = filterText;
-    filterEl.filterProps = filterProps;
+    filterEl.filterProps = effectiveFilterProps;
     this.filterAndUpdateData();
   }
 
@@ -757,7 +765,7 @@ export class List
       label: item.label,
       description: item.description,
       metadata: item.metadata,
-      value: item.value,
+      el: item,
     }));
   }
 
@@ -960,7 +968,7 @@ export class List
       hasFilterActionsStart,
       hasFilterActionsEnd,
       hasFilterNoResults,
-      filterProps,
+      effectiveFilterProps,
     } = this;
     return (
       <InteractiveContainer disabled={this.disabled}>
@@ -992,7 +1000,7 @@ export class List
                       <calcite-filter
                         ariaLabel={filterPlaceholder}
                         disabled={disabled}
-                        filterProps={filterProps}
+                        filterProps={effectiveFilterProps}
                         items={dataForFilter}
                         oncalciteFilterChange={this.handleFilterChange}
                         placeholder={filterPlaceholder}