+++ packages/calcite-components-react/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.0-next.18](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.17...@esri/calcite-components-react@3.0.0-next.18) (2024-11-20)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.17](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.16...@esri/calcite-components-react@3.0.0-next.17) (2024-11-20)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.16](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.15...@esri/calcite-components-react@3.0.0-next.16) (2024-11-19)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.15](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.14...@esri/calcite-components-react@3.0.0-next.15) (2024-11-18)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.14](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.13...@esri/calcite-components-react@3.0.0-next.14) (2024-11-18)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.13](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.12...@esri/calcite-components-react@3.0.0-next.13) (2024-11-18)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.12](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.11...@esri/calcite-components-react@3.0.0-next.12) (2024-11-15)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.11](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.10...@esri/calcite-components-react@3.0.0-next.11) (2024-11-15)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.10](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.9...@esri/calcite-components-react@3.0.0-next.10) (2024-11-15)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.9](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.8...@esri/calcite-components-react@3.0.0-next.9) (2024-11-15)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [3.0.0-next.8](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.7...@esri/calcite-components-react@3.0.0-next.8) (2024-11-15)
 
 ### ⚠ BREAKING CHANGES